Reduce complexity of edit comment endpoint [#162840142]  On branch reduce-complexity  Changes to be committed: 	modified:   .coverage 	modified:   api/app.py 	new file:   api/routes/edit_incident.py 	modified:   api/routes/incidents.py
