+++ api/app.py
@@ -6,6 +6,7 @@ from api.helpers.responses import supported_end_points
 from api.routes.admin import admin_flags_bp
 from api.routes.auth import users_bp
 from api.routes.incidents import red_flags_bp
+from api.routes.edit_incident import edit_red_flags_bp
 from config import Config
 
 
@@ -29,6 +30,7 @@ def create_app(config="None"):
     app.config.from_object(Config)
     app.register_blueprint(users_bp)
     app.register_blueprint(red_flags_bp)
+    app.register_blueprint(edit_red_flags_bp)
     app.register_blueprint(admin_flags_bp)
 
     return app
